// src/app/models/turn.ts
export interface Turn {
    id: number;
    code: string;
    startHour: Date;
    endHour: Date;
    status: boolean;
    userId: number;
}

/**
 * Interface turnos para modulo de nomina
 */
/*export interface WorkloadTurnDTO {
    turnId: number;
    turnCode: string;
    startHour: string;
    endHour: string;
    sampleTakeDate: string;
    checked: boolean;
}*/